support python < 3.5
